Correct EOF when receiving valid ack packets with no properties lenght

+++ src/commonMain/kotlin/socket/streams/ByteArrayInputStream.kt
@@ -6,7 +6,7 @@ class ByteArrayInputStream(private val array: UByteArray) : InputStream {
 
     override fun read(): UByte {
         return if (position < array.size)
-            array[position++].toUByte()
+            array[position++]
         else
             throw EOFException()
     }
